Remade the drag system
The dragging modification system has been changed so that you need to drive into the modules before they are draggable.

There is also no a module swap feature where if you drag a module on top of another module, they will swap positions around your ship

And there is also a single inventory slot on the left of the screen for whatever we want to do with it.