Healthbar V12 Part 2 - Existential Dread
Taking a complete 180 in the opposite direction but bloating my script anyway!

The Problem: Who actually deals with health?
There are two paths we can take here:
A) Module/Player scripts depend on Healthbar.cs to store health
B) Healthbar.cs depends on Player/Module scripts to store health

If we go A), it means we wont have to write a lot of code in modules, possibly save reuse of code

But if we go B), it means we can do so much more with the Healthbar,
as manually setting the Healthbar's values means it will be a lot easier to expand on with features
like a GUI (health?) bar for 550 health Boss Battle, or a GUI bar 60 Aura etc.

What I am doing right now is an indecision: I am preparing for BOTH possibilities, including a method for manually setting health and a method for manually setting maxHealth. Although the current graphic is a bunch of heart containers, it is actually just a progress bar under the hood (i've probably mentioned this before)
So maybe my efforts are futile here but my ACTUAL decision as far as integration is concerned is I am going to go back to option A), but keep the code for option B) just in case.
It is still very ugly, and now my healthbar script is even larger and MORE unreadable but that is ok because noone has to see it! I dunno, i have to save this commit. its a terrible commit and im probably gonna undo as i havent even done much yet but thats all from me bye